+++ src/Face.hpp
@@ -68,7 +68,7 @@ public:
     const Face &_face;
 
     /*! @brief Vertex the iterator is currently pointing to. */
-    unsigned int _index;
+    size_t _index;
 
   public:
     /**
@@ -77,7 +77,7 @@ public:
      * @param face Underlying Face object.
      * @param index Vertex the iterator is currently pointing to.
      */
-    inline Vertices(const Face &face, unsigned int index)
+    inline Vertices(const Face &face, size_t index)
         : _face(face), _index(index) {}
 
     /**
